Add reduce/aggregation trees (#590)
* Add reduce trees


* Remove task in TODO


* Remove log statements
